Fix for rebase issue.

+++ lib/basic/atca_helpers.h
@@ -53,7 +53,7 @@ ATCA_STATUS atcab_printbin_label(const char* label, uint8_t* binary, size_t bin_
 #endif
 
 
-ATCA_STATUS packHex(const char* ascii_hex, int ascii_hex_len, char* packed_hex, int* packed_len);
+ATCA_STATUS packHex(const char* ascii_hex, size_t ascii_hex_len, char* packed_hex, size_t* packed_len);
 bool isDigit(char c);
 bool isWhiteSpace(char c);
 bool isAlpha(char c);
